Move SocialController
